requirements.txt
numpy==1.20.1
opencv-contrib-python==4.5.1.48
Pillow==8.1.2